refactor - to separate macros and functions in Main.cpp to config/* and util/path.hpp
