+++ include/config/initial.hpp
@@ -1,0 +1,7 @@
+/*****************************
+ * config/initial.hpp
+ *****************************/
+
+#pragma once
+
+#define STATIC_CONSTEXPR static constexpr
