+++ include/path.hpp
@@ -1,0 +1,10 @@
+/*****************************
+ * path
+ *****************************/
+
+#include <vector>
+#include <string>
+
+#pragma once
+
+void recursive_scan_directory( std::string const &, std::vector<std::string> & );
