+++ include/util/path.hpp
@@ -1,0 +1,10 @@
+/*****************************
+ * util/path
+ *****************************/
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+void recursive_scan_directory( std::string const &scanning_path, std::vector<std::string> &paths );
